+++ source/Cosmos.Core/IOGroup/Mouse.cs
@@ -13,10 +13,10 @@ namespace Cosmos.Core.IOGroup
         /// <summary>
         /// Data port.
         /// </summary>
-        public readonly ushort p60 = 0x60;
+        public readonly int p60 = 0x60;
         /// <summary>
         /// Indicator port, used to tell if data came from keyboard or mouse.
         /// </summary>
-        public readonly ushort p64 = 0x64;
+        public readonly int p64 = 0x64;
     }
 }
